Consolidate refactored modules to reduce PR size
Optimization pass to remove over-abstraction:

1. Context Plumbing (-350 lines)
   - Merged narrativeGenerationContext.ts into contextBuilder.ts
   - Removed narrativeContextGateway.ts wrapper layer
   - ContextBuilder now reads stores directly

2. Prompt Composer (-200 lines)
   - Flattened promptComposer/ tree into single promptEnhancers.ts
   - Merged all small enhancers (Lore, Goal, Inventory, Item, Tone)
   - Kept data-driven configuration without boilerplate

3. Post-Processing (-250 lines)
   - Consolidated 5 files into responseProcessor.ts
   - Merged LoreUpdater, InventoryManager into ResponseProcessor
   - Reuses existing language complexity helpers

4. Generation Pipeline (-480 lines)
   - Inlined pipeline orchestration back into NarrativeGenerator
   - Kept modular helpers without abstraction overhead
   - NarrativeGenerator now ~540 lines (vs original ~2000)

Total reduction: ~1,280 lines from first refactoring pass
Net change: +683/−1,815 (down from +2,498/−1,815)

All public methods remain unchanged - backward compatible.
