--- src/lib/ai/postProcessing/index.ts
@@ -1,10 +1,0 @@
-/**
- * Post-Processing Modules
- *
- * Export all post-processing modules for easy import
- */
-
-export { ResponseParser } from './ResponseParser';
-export { LoreUpdater } from './LoreUpdater';
-export { InventoryManager } from './InventoryManager';
-export { LanguageComplexityEnforcer } from './LanguageComplexityEnforcer';
