--- src/lib/ai/promptComposer/enhancers/index.ts
@@ -1,12 +1,0 @@
-/**
- * Prompt Enhancers
- *
- * Export all available enhancers for easy import
- */
-
-export { ToneSettingsEnhancer } from './ToneSettingsEnhancer';
-export { LoreEnhancer } from './LoreEnhancer';
-export { GoalContextEnhancer } from './GoalContextEnhancer';
-export { PersonalizationEnhancer } from './PersonalizationEnhancer';
-export { InventoryEnhancer } from './InventoryEnhancer';
-export { ItemAcquisitionEnhancer } from './ItemAcquisitionEnhancer';
